refactor: make search bar folder name consistent with its content
